organize types file, refactor useAbilityScore hook
