+++ src/components/Navbar/NavDrawer.tsx
@@ -1,15 +1,15 @@
-import { useState } from 'react';
+import ArrowCircleRightIcon from '@mui/icons-material/ArrowCircleRight';
 import Box from '@mui/material/Box';
-import Drawer from '@mui/material/Drawer';
 import Button from '@mui/material/Button';
+import Drawer from '@mui/material/Drawer';
 import List from '@mui/material/List';
 import ListItem from '@mui/material/ListItem';
 import ListItemButton from '@mui/material/ListItemButton';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
-import ArrowCircleRightIcon from '@mui/icons-material/ArrowCircleRight';
-import { useCategories } from '../../hooks/useCategories';
+import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+import { useCategories } from '../../hooks/useMemo/useCategories';
 
 export default function NavDrawer() {
   const [open, setOpen] = useState(false);
